refactor(server): split routers/graph.py into a graph/ sub-router package
Break the 1355-line graph router into per-concern sub-routers behind a
graph/__init__.py façade that assembles `router` exactly as before (zero
caller churn — app.py still mounts `graph.router`).

- _common.py: `with_repo` FastAPI dependency collapsing the 14× repeated
  fetch-repo → 404 guard; `_edge_response`, `_escape_like`, name helpers.
- signals.py: NodeSignals + `_collect_node_signals` + a generic
  `_node_to_response(n, sig, Cls, **extra)` that collapses the
  `_to_graph_node`/`_to_dead_node`/`_to_hot_node` 16-field triples.
- sub-routers: module_graph, neighborhoods (ego/entry-points/dead/hot),
  community (architecture + communities), intelligence (metrics/callers/
  flows), path, full_graph (export + search).

Behavior-preserving: route paths, ordering (single-segment catch-all last),
404 details, and response shapes unchanged. tests/unit/server (259) green;
ruff clean; no net-new mypy errors (the one max(key=dict.get) note is a
verbatim move from the old graph.py:299).
